content/html/content/test/test_bug573969.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=573969
-->
<head>
  <title>Test for Bug 573969</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=573969">Mozilla Bug 573969</a>
<p id="display"></p>
<div id="content" style="display: none">
  <xmp id='x'></xmp>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 573969 **/

var testData = [
  '<div>foo</div>',
  '&lt;div&gt;&lt;/div&gt;',
];

var x = document.getElementById('x');

for each (v in testData) {
  x.innerHTML = v;
  is(x.innerHTML, v, "innerHTML value should not be escaped");
}

</script>
</pre>
</body>
</html>
